+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Improve types [#160](https://github.com/CartoDB/carto-react/pull/160)
+- Adapt Legend widget to work without children [#161](https://github.com/CartoDB/carto-react/pull/161)
 - Update deck.gl version to 8.5.0 [#162](https://github.com/CartoDB/carto-react/pull/162)
 - Fix Material-UI warnings on justify property [#162](https://github.com/CartoDB/carto-react/pull/162)